Refactor the application, introduce MainScene, ControlPanel, and UrlHandler objects
