+++ .editorconfig
@@ -1,0 +1,12 @@
+root = true
+
+# Unix-style newlines with a newline ending every file
+[*]
+indent_style = tab
+charset = utf-8
+end_of_line = lf
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.md]
+trim_trailing_whitespace = false
