--- src/models/README.md
@@ -1,1 +1,0 @@
-Your models here.
