--- src/sounds/README.md
@@ -1,1 +1,0 @@
-Your sounds here.
